Heredoc fully working with pipes. Exec almost there
