+++ Makefile
@@ -46,6 +46,8 @@ SRC 		:= 	main/minishell.c 				\
 				executor/executor_utils3.c		\
 				executor/executor_utils4.c		\
 				executor/executor_utils5.c		\
+				executor/executor_utils6.c		\
+				executor/executor_utils7.c		\
 				executor/piping.c 				\
 				executor/piping_utils1.c		\
 				executor/redirection.c 			\
