+++ src/executor/executor.c
@@ -6,7 +6,7 @@
 /*   By: hmunoz-g <hmunoz-g@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/25 11:42:26 by hmunoz-g          #+#    #+#             */
-/*   Updated: 2024/12/28 12:24:14 by hmunoz-g         ###   ########.fr       */
+/*   Updated: 2024/12/30 13:15:06 by hmunoz-g         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,59 +36,29 @@ int	ms_handle_system_cmd(t_ms *ms, char **env)
 	return (-1);
 }
 
-// Main execution function
-int	ms_exec_command(t_ms *ms, char **env)
-{
-	if (!ms->cmd_args)
-		return (ms_error_handler(ms, "Error: Mem alloc failed", 1), 1);
-	if (ms_has_redirection(ms))
-	{
-		if (ms_redirection(ms) == -1)
-		{
-			if (ms_is_builtin(ms->filt_args[0]))
-				return (1);
-			exit (1);
-		}
-	}
-	if (ms_is_builtin(ms->cmd_args[0]))
-	{
-		if (ms_reroute_builtins(ms, env))
-			return (1);
-	}
-	else if (ms_handle_system_cmd(ms, env) == -1)
-	{
-		ft_free(ms->cmd_args);
-		ft_free(ms->filt_args);
-		return (1);
-	}
-	return (1);
-}
-
-/*
-Executor hub.
-*/
 void	ms_process_command(t_ms *ms, char **env, int i)
 {
 	pid_t	pid;
 	int		saved_fds[3];
 
+	if (ms_has_heredoc(ms) && ms_handle_heredoc_setup(ms) == -1)
+		return (ms_error_handler(ms, "Error: Failed to redir input", 0));
 	if (ms->filt_args[0] && ms_is_builtin(ms->filt_args[0]) && !ms->pipe_count)
 	{
 		ms_save_std_fds(saved_fds);
+		if (ms_has_redirection(ms))
+			ms_redirection(ms);
 		ms_exec_command(ms, env);
 		ms_restore_std_fds(saved_fds);
+		ms_cleanup_heredoc(ms);
 	}
 	else
 	{
 		pid = fork();
 		if (pid == 0)
-		{
-			ms_setup_child_pipes(ms->pipe_fds, i, ms->pipe_count);
-			ms_close_child_pipes(ms->pipe_fds, ms->pipe_count);
-			if (ms_exec_command(ms, env) != 0)
-				exit (1);
-			exit(0);
-		}
+			ms_handle_child_process(ms, env, i);
+		else if (pid > 0)
+			ms_handle_parent_process(ms);
 	}
 }
 
@@ -112,7 +82,11 @@ void	ms_executor(t_ms *ms)
 	ms_initialize_execution(ms, &env);
 	i = -1;
 	while (++i < ft_array_count(ms->exec_chunks))
+	{
 		ms_execute_chunk(ms, env, i);
+		if (i < ft_array_count(ms->exec_chunks) - 1)
+			ms_close_used_pipes(ms->pipe_fds, i);
+	}
 	ms_close_parent_pipes(ms->pipe_fds, ms->pipe_count);
 	ms_wait_children(ft_array_count(ms->exec_chunks));
 	ms_executor_cleanup(ms, env);
