+++ src/executor/executor_utils2.c
@@ -6,53 +6,76 @@
 /*   By: hmunoz-g <hmunoz-g@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/25 11:42:26 by hmunoz-g          #+#    #+#             */
-/*   Updated: 2024/12/20 15:30:41 by hmunoz-g         ###   ########.fr       */
+/*   Updated: 2024/12/30 13:16:48 by hmunoz-g         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 /*
-This file contains helper/Flow-control functions to build exec components:
-	-ms_env_to_aray
-	-ms_rebuild_env
+This file contains 
 */
 
 #include "../../includes/minishell.h"
 
-char	**ms_env_to_array(t_ms *ms, char **arr)
+// Main execution function
+int	ms_exec_command(t_ms *ms, char **env)
 {
-	t_list	*current;
-	int		i;
-
-	(void)ms;
-	current = ms->ms_env;
-	i = 0;
-	while (current)
+	if (!ms->cmd_args)
+		return (ms_error_handler(ms, "Error: Mem alloc failed", 1), 1);
+	if (ms_has_redirection(ms))
 	{
-		if (ft_strchr(current->content, '='))
+		if (ms_redirection(ms) == -1)
 		{
-			arr[i] = ft_strdup((char *)current->content);
-			if (!arr[i])
-			{
-				ft_free(arr);
-				return (NULL);
-			}
-			i++;
+			if (ms_is_builtin(ms->filt_args[0]))
+				return (1);
+			exit(1);
 		}
-		current = current->next;
 	}
-	arr[i] = NULL;
-	return (arr);
+	if (ms_is_builtin(ms->cmd_args[0]))
+	{
+		if (ms_reroute_builtins(ms, env))
+			return (1);
+	}
+	else if (ms_handle_system_cmd(ms, env) == -1)
+	{
+		ft_free(ms->cmd_args);
+		ft_free(ms->filt_args);
+		return (1);
+	}
+	return (1);
 }
 
-char	**ms_rebuild_env(t_ms *ms)
+/*
+Executor hub.
+*/
+void	ms_handle_child_process(t_ms *ms, char **env, int i)
 {
-	char	**arr;
+	if (ms->heredoc_fd != -1 && dup2(ms->heredoc_fd, STDIN_FILENO) == -1)
+	{
+		ms_error_handler(ms, "Heredoc dup2 failed", 0);
+		exit(1);
+	}
+	ms_setup_child_pipes(ms, i, ms->pipe_count);
+	if (ms_has_redirection(ms) && !ms_has_heredoc(ms)
+		&& ms_redirection(ms) == -1)
+		exit(1);
+	if (ms_exec_command(ms, env) != 0)
+		exit(1);
+	exit(0);
+}
 
-	arr = (char **)malloc(sizeof(char *) * (ft_lstsize(ms->ms_env) + 1));
-	if (!arr)
-		return (ms_error_handler(ms, "Error: Mem alloc failed", 1), NULL);
-	arr = ms_env_to_array(ms, arr);
-	if (!arr)
-		return (ms_error_handler(ms, "Failed to prepare environment", 0), NULL);
-	return (arr);
+void	ms_handle_parent_process(t_ms *ms)
+{
+	if (ms->heredoc_fd != -1)
+	{
+		close(ms->heredoc_fd);
+		ms->heredoc_fd = -1;
+	}
+}
+
+void	ms_close_used_pipes(int **pipe_fds, int i)
+{
+	if (i > 0)
+		close(pipe_fds[i - 1][0]);
+	if (pipe_fds[i])
+		close(pipe_fds[i][1]);
 }
